assignment_I_&_II/requirements.txt
pandas==1.4.3
uvicorn==0.18.2
fastapi==0.79.0
sqlalchemy==1.4.39
scikit-learn==1.1.1
pytest==7.1.2
requests==2.28.1
pyyaml==5.4.1
psycopg2==2.9.3